# openapi-default.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/verification/repositories": {
      "get": {
        "operationId": "AppController_getConstantsRepositories",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/organization": {
      "post": {
        "operationId": "OrganizationController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        },
        "tags": [
          "organization"
        ]
      },
      "get": {
        "operationId": "OrganizationController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Organization Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        },
        "tags": [
          "organization"
        ]
      }
    },
    "/organization/{id}": {
      "get": {
        "operationId": "OrganizationController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "organization"
        ]
      },
      "patch": {
        "operationId": "OrganizationController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "organization"
        ]
      },
      "delete": {
        "operationId": "OrganizationController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "organization"
        ]
      }
    },
    "/repository/{id}": {
      "get": {
        "operationId": "RepositoryController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/repository/download/{id}": {
      "get": {
        "operationId": "RepositoryController_getRepositoryDownloadMetrics",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "info": {
    "title": "Challenge Backend",
    "description": "De Una! Pichincha",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "organization",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "CreateOrganizationDto": {
        "type": "object",
        "properties": {}
      },
      "Organization": {
        "type": "object",
        "properties": {}
      },
      "UpdateOrganizationDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}